requirements.txt
# Web Server & Sockets
Flask==3.1.3
Flask-SocketIO==5.6.1
python-socketio==5.16.1
eventlet==0.41.0
Werkzeug==3.1.8

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.49

# AI & Audio Processing
vosk==0.3.45
pydub==0.25.1
librosa==0.11.0
soundfile==0.13.1
google-generativeai

# Japanese Language & Scoring Logic
jaconv==0.5.0
Levenshtein==0.27.3
python-Levenshtein==0.27.3
RapidFuzz==3.14.5

# Data Handling
pandas==3.0.2
numpy==2.4.4

# Utilities
requests==2.33.1
tqdm==4.67.3